fix: modify field type

+++ OTC/broker/orderProcessor.py
@@ -5,7 +5,7 @@ def add_buy_order(broker: Broker,
                   buyer_id: int,
                   commodity_name: str,
                   buy_vol: float,
-                  price: float,
+                  price: int,
                   order_type: str):
     order_id = broker.buy_order_id.get(commodity_name, 1)
     broker.buy_order_id[commodity_name] = broker.buy_order_id.get(commodity_name, 1) + 1
@@ -24,7 +24,7 @@ def add_sell_order(broker: Broker,
                    seller_id: int,
                    commodity_name: str,
                    sell_vol: float,
-                   price: float,
+                   price: int,
                    order_type: str):
     order_id = broker.sell_order_id.get(commodity_name, 1)
     broker.sell_order_id[commodity_name] = broker.sell_order_id.get(commodity_name, 1) + 1
